refactor: reorganize app structure using route groups
