+++ src/app/(auth)/join/page.tsx
