+++ src/app/(main)/layout.tsx
@@ -1,0 +1,9 @@
+import { NavLayout } from '@/components/NavLayout/NavLayout';
+
+export default function MainLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return <NavLayout>{children}</NavLayout>;
+}
